missing then

+++ deploy/scripts/install_workloadzone.sh
@@ -159,10 +159,16 @@ if [ -n "${temp}" ]; then
     echo "#                                                                                       #"
     echo "#########################################################################################"
     echo ""
-    rm stdout.az
+    if [ -f stdout.az ]
+    then
+        rm stdout.az
+    fi
     exit -1
 else
-    rm stdout.az
+    if [ -f stdout.az ]
+    then
+        rm stdout.az
+    fi
 fi
 account_set=0
